+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/notification/dto/NotificationTokenAssignRequest.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.notification.dto.request;
+package com.startingblue.fourtooncookie.notification.dto;
 
 public record NotificationTokenAssignRequest(String notificationToken) {
 }
